Calculator Application
The Calculator application is a one-stop android app which handles:

basic mathematic operations like addition, subtraction,
multiplication, and division between two operands

finding the power of an operand and toggling the sign of an operand.

memory operations like adding a value to memory, subtracting a value
from memory, recalling what is stored in memory and clearing the
memory content.